Update Program.cs
added some code to write to a text file, make sure to get your data from the text file before running the program again as it will overwrite the text file.  Also you have to create the text file and leave it empty before running the program for the first time.